% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spde.R
\name{create_spde}
\alias{create_spde}
\title{Create a mesh and SPDE}
\usage{
create_spde(spatial_grid, config_sp)
}
\arguments{
\item{spatial_grid}{A sfc POINT object representing the full spatial grid/}

\item{config_sp}{A list that should contains the following parameters:
\itemize{
\item alpha: the smoothness parameter of the Matern covariance function
\item kappa: the range parameter of the Matern covariance function
\item variance: the variance parameter of the Matern covariance function
}}
}
\value{
a list containing the mesh, spde, Q, and A
}
\description{
Create a mesh and SPDE
}
\details{
This function creates a mesh and SPDE for a given spatial grid and
config_spuration.  It is primarily a wrapper for the create_spde_mesh and
create_spde_matern functions.
}
\examples{
library(sf)
library(INLA)
config_sp <- list(crs=4326, seed = 123)
spatial_domain <- st_geometry(
  st_multipoint(
    x = rbind(
      c(0, -10),
      c(3, -10),
      c(10, -20),
      c(1, -21),
      c(2, -16),
      c(0, -10)
    )
  )
) |>
  st_set_crs(config_sp$crs) |>
  st_cast("POLYGON")
set.seed(config_sp$seed)
spatial_grid <- spatial_domain |>
  st_set_crs(NA) |>
  st_sample(size = 10000, type = "regular") |>
  st_set_crs(config_sp$crs)
config_sp <- list(alpha = 2, kappa = 1, variance = 1)
matern_projection <- create_spde(spatial_grid, config_sp)
}
\author{
Murray
}
