% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spde_matern.R
\name{create_spde_matern}
\alias{create_spde_matern}
\title{Create SPDE Matern Components}
\usage{
create_spde_matern(spatial_grid, mesh, config)
}
\arguments{
\item{spatial_grid}{An sf object containing the spatial grid.}

\item{mesh}{An INLA mesh object.}

\item{config}{A list with SPDE parameters, including \code{alpha}, \code{variance},
and \code{kappa}.}
}
\value{
A list containing:
\item{spde}{The INLA SPDE Matern model.}
\item{Q}{The precision matrix derived from SPDE parameters.}
\item{A}{The projection matrix mapping mesh nodes to grid points.}
}
\description{
Constructs the SPDE Matern model components (spde object, precision matrix,
and projection matrix) for a given spatial grid and INLA mesh.
}
\author{
Murray
}
