% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spde_mesh.R
\name{create_spde_mesh}
\alias{create_spde_mesh}
\title{Create SPDE Mesh}
\usage{
create_spde_mesh(spatial_grid, config)
}
\arguments{
\item{spatial_grid}{An sf object representing the spatial grid from which
mesh nodes will be generated.}

\item{config}{A list containing SPDE parameters (e.g., \code{alpha}, \code{kappa})
used to define mesh spacing and scaling.}
}
\value{
An \code{inla.mesh} object representing the constructed SPDE mesh.
}
\description{
Create an SPDE Mesh
}
\details{
Constructs a 2D INLA SPDE mesh from the convex hull of an input spatial grid,
using mesh parameters derived from the model configuration.
}
\author{
Murray
}
