% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_design_large_scale_fixed.R
\name{sampling_design_large_scale_fixed}
\alias{sampling_design_large_scale_fixed}
\title{Large‐scale fixed sampling design}
\usage{
sampling_design_large_scale_fixed(data_reefs_pts_sf, config_lrge)
}
\arguments{
\item{data_reefs_pts_sf}{An \code{sf} object representing the full field.
It must contain:
\itemize{
\item \code{Year}: numeric year
\item \code{Reef}: unique reef ID
\item \code{HCC}, \code{SC}, \code{MA}: cover values (logit scale)
\item \code{geometry}: spatial geometry
}}

\item{config_lrge}{A list with:
\itemize{
\item \code{n_locs}: number of reefs (locations) to select
\item \code{n_sites}: number of sites per selected reef
\item \code{seed}: random seed
}}
}
\value{
An \code{sf} object representing the large-scale sampling design.
Cover values remain on the logit scale.
}
\description{
Large‐scale fixed sampling design
}
\details{
Selects a set number of reef locations and then randomly selects a
fixed number of sites within each selected location.
}
\examples{
config_lrge <- list(n_locs = 25, n_sites = 2, seed = 123)
benthos_fixed_locs_sf <- sampling_design_large_scale_fixed(
  benthos_reefs_pts, config_lrge
)

}
\author{
Murray
}
