% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grid_sfc_to_df.R
\name{spatial_grid_sfc_to_df}
\alias{spatial_grid_sfc_to_df}
\title{Convert sf Geometry to Data Frame}
\usage{
spatial_grid_sfc_to_df(spatial_grid)
}
\arguments{
\item{spatial_grid}{An sf object containing point geometries from which
coordinates will be extracted.}
}
\value{
A data frame with columns \code{Longitude} and \code{Latitude}, ordered
lexicographically.
}
\description{
Convert sf Geometry to a Coordinate Data Frame
}
\details{
Extracts point coordinates from an sf geometry column (\code{sfc}) and returns
them as a tidy data frame with renamed longitude and latitude columns.
}
\author{
Murray
}
