% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_soft_coral_cover.R
\name{baseline_soft_coral_cover}
\alias{baseline_soft_coral_cover}
\title{Baseline Soft Coral Cover}
\usage{
baseline_soft_coral_cover(spatial_grid, spde, config_sp)
}
\arguments{
\item{spatial_grid}{An \code{sfc_POINT} object representing the full spatial grid.}

\item{spde}{A list containing the SPDE mesh, SPDE object, precision matrix \code{Q},
and projection matrix \code{A}.}

\item{config_sp}{A list containing config_spuration parameters including \code{years}.}

\item{cover_range}{Numeric vector of length 2 defining the broad-scale range
of soft coral cover on the link scale. Values must be >0 and <1. Default \code{c(0.01, 0.1)}.}
}
\value{
A list with:
\itemize{
\item \code{baseline_sample_sc} – baseline soft coral cover sample
\item \code{baseline_effects_sc} – matrix of baseline effects
\item \code{baseline_pts_sample_sc} – projected baseline onto the spatial grid
\item \code{baseline_pts_effects_sc} – long-format data frame suitable for plotting
}
}
\description{
Baseline Soft Coral Cover
}
\details{
Calculates the baseline spatial pattern of soft coral cover prior to sampling.
The pattern is defined as a simple sine wave (applied to centered latitudes)
and projected onto the SPDE grid. Values are on the link (logit) scale.
}
\author{
Murray
}
