% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reefs.R
\name{generate_reefs}
\alias{generate_reefs}
\title{Generate simulated reefs}
\usage{
generate_reefs(simulated_patch_sf, config)
}
\arguments{
\item{simulated_patch_sf}{A simulated patch as an sf object as generated by the function
\code{generate_patch}}

\item{config}{A list that contains the parameters for filtering the field to patches.
The list should contain the following parameters:
\itemize{
\item reef_width: a numeric representing the half the width of a ribbon
representing the reef that is centered on the outline of the patch
}}
}
\value{
a list containing:
\itemize{
\item simulated_reefs_sf: a sfc multipolygon object that represents the simulated reefs
\item simulated_reefs_poly_sf: an sf object that represents the simulated reefs
}
}
\description{
Generate simulated reefs
}
\details{
Generate simulated reefs from a simulated patches.  A "reef" is a
special type of patch that is created by buffering the outline of a
patch to create a ribbon that is centered on the outline of the patch.
The width of the ribbon is determined by the parameter reef_width.
In this way, the resultin object more closely resembles a coral reef
which tends to be an irregular ring or semi-ring shaped polygon with
the coral growing mostly around the outside edge surrounding a
central sandy lagoon.
}
\examples{
library(sf)
library(gstat)
library(ggplot2)
config <- list(
 seed = 1,
 crs = 4326,
 model = "Exp",
 psill = 1,
 range = 15,
 nugget = 0,
 patch_threshold = 1.75,
 reef_width = 0.01
)
spatial_domain <- st_geometry(
  st_multipoint(
    x = rbind(
      c(0, -10),
      c(3, -10),
      c(10, -20),
      c(1, -21),
      c(2, -16),
      c(0, -10)
    )
  )
) |>
  st_set_crs(config$crs) |>
  st_cast("POLYGON")
set.seed(config$seed)
spatial_grid <- spatial_domain |>
  st_set_crs(NA) |>
  st_sample(size = 10000, type = "regular") |>
  st_set_crs(config$crs)
simulated_field <- generate_field(spatial_grid, config)
simulated_patches <- generate_patches(simulated_field, config)
simulated_reefs <- generate_reefs(simulated_patches, config)
ggplot() +
   geom_sf(data = simulated_reefs$simulated_reefs_sf, fill = "lightblue") +
   coord_sf(xlim = c(1.8, 2.3), ylim = c(-20.4, -19.9)) 
}
\author{
Murray
}
