% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_design_large_scale_fixed.R
\name{sampling_design_fine_scale_fixed}
\alias{sampling_design_fine_scale_fixed}
\title{Fine scale fixed sampling design}
\usage{
sampling_design_fine_scale_fixed(data_fixed_locs_sf, config)
}
\arguments{
\item{data_fixed_locs_sf}{A sf containing the large scale sampling design with:
\itemize{
\item Reef: a unique reef id
\item Site: a unique site id
\item geometry: sf geometry
}}

\item{config}{A list containing:
\itemize{
\item Number_of_transects_per_site: number of transects per site
\item Depths: number of depths
\item Number_of_frames_per_transect: number of frames per transect
\item Points_per_frame: number of points per frame
\item hcc_site_sigma: variability (on logit scale) in Sites within Locations
\item hcc_transect_sigma: variability (on logit scale) in Transects within Sites
\item hcc_sigma: random noise (on a logit scale)
\item sc_site_sigma: variability (on a logit scale) in Sites within Locations
\item sc_transect_sigma: variability (on a logit scale) in Transects within Sites
\item sc_sigma: random noise (on a logit scale)
\item ma_site_sigma: variability (on a logit scale) in Sites within Locations
\item ma_transect_sigma: variability (on a logit scale in Transects within Sites
\item ma_sigma: random noise (on a logit scale
}}
}
\value{
A sf object denoting the fine scale sampling design.  Cover values are on natural percentage cover scale
}
\description{
Fine scale fixed sampling design
}
\details{
Given a large scale sampling design and a number of parametes,
create a spatial hierarchy of transects per site, photo-frames
per transect and points per frame.
}
\examples{
library(sf)
library(stars)
library(gstat)
library(INLA)
config <- list(
  seed = 1,
  crs = 4326,
  model = "Exp",
  psill = 1,
  range = 15,
  nugget = 0,
  alpha = 2,
  kappa = 1,
  variance = 1,
  patch_threshold = 1.75,
  reef_width = 0.01,
  years = 1:12,
  dhw_weight = 0.5,
  cyc_weight = 0.4,
  other_weight = 0.1,
  hcc_cover_range = c(0.1, 0.7),
  hcc_growth = 0.3,
  sc_cover_range = c(0.01, 0.1),
  sc_growth =  0.3
)
spatial_domain <- st_geometry(
  st_multipoint(
    x = rbind(
      c(0, -10),
      c(3, -10),
      c(10, -20),
      c(1, -21),
      c(2, -16),
      c(0, -10)
    )
  )
) |>
  st_set_crs(config$crs) |>
  st_cast("POLYGON")
set.seed(config$seed)
spatial_grid <- spatial_domain |>
  st_set_crs(NA) |>
  st_sample(size = 10000, type = "regular") |>
  st_set_crs(config$crs)
benthos_reefs_pts <- create_synthetic_reef_landscape(spatial_grid, config)
config <- list(n_locs = 25, n_sites = 2, seed = 123)
benthos_fixed_locs_sf <- sampling_design_large_scale_fixed(benthos_reefs_pts, config)
config <- list(
  Number_of_transects_per_site = 5,
  Depths = 2,
  Number_of_frames_per_transect = 100,
  Points_per_frame = 5,
  ## Note, the following are on the link scale
  hcc_site_sigma = 0.5, # variability in Sites within Locations
  hcc_transect_sigma = 0.2, # variability in Transects within Sites
  hcc_sigma = 0.1, # random noise
  sc_site_sigma = 0.05, # variability in Sites within Locations
  sc_transect_sigma = 0.02, # variability in Transects within Sites
  sc_sigma = 0.01, # random noise
  ma_site_sigma = 0.5, # variability in Sites within Locations
  ma_transect_sigma = 0.2, # variability in Transects within Sites
  ma_sigma = 0.1 # random noise
)
benthos_fixed_locs_obs <- sampling_design_fine_scale_fixed(benthos_fixed_locs_sf, config)
}
\author{
Murray
}
