% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_design_large_scale_fixed.R
\name{sampling_design_large_scale_fixed}
\alias{sampling_design_large_scale_fixed}
\title{Large scale fixed sampling design}
\usage{
sampling_design_large_scale_fixed(data_reefs_pts_sf, config)
}
\arguments{
\item{data_reefs_pts_sf}{A sf containing the full field with:
\itemize{
\item Year: a numeric representation of years in the temporal domain
\item Reef: a unique reef id
\item HCC: hard coral cover (on logit scale)
\item SC: soft coral cover (on logit scale)
\item MA: macroalgae cover (on logit scale)
\item geometry: sf geometry
}}

\item{config}{A list containing:
\itemize{
\item n_locs: number of reefs (locations)
\item n_sites: number of sites per location
\item seed: random seed
}}
}
\value{
A sf object denoting the large scale sampling design.  Cover values are on a logit scale.
}
\description{
Large scale fixed sampling design
}
\details{
Given a number of locations and a number of sites per location,
select the locations from the full field (domain) and then
randomly select the sites within those locations.
}
\examples{
library(sf)
library(stars)
library(gstat)
library(INLA)
config <- list(
  seed = 1,
  crs = 4326,
  model = "Exp",
  psill = 1,
  range = 15,
  nugget = 0,
  alpha = 2,
  kappa = 1,
  variance = 1,
  patch_threshold = 1.75,
  reef_width = 0.01,
  years = 1:12,
  dhw_weight = 0.5,
  cyc_weight = 0.4,
  other_weight = 0.1,
  hcc_growth = 0.3,
  sc_growth =  0.3
)
spatial_domain <- st_geometry(
  st_multipoint(
    x = rbind(
      c(0, -10),
      c(3, -10),
      c(10, -20),
      c(1, -21),
      c(2, -16),
      c(0, -10)
    )
  )
) |>
  st_set_crs(config$crs) |>
  st_cast("POLYGON")
set.seed(config$seed)
spatial_grid <- spatial_domain |>
  st_set_crs(NA) |>
  st_sample(size = 10000, type = "regular") |>
  st_set_crs(config$crs)
benthos_reefs_pts <- create_synthetic_reef_landscape(spatial_grid, config)
config <- list(n_locs = 25, n_sites = 2, seed = 123)
benthos_fixed_locs_sf <- sampling_design_large_scale_fixed(benthos_reefs_pts, config)
}
\author{
Murray
}
