% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_field_hcc.R
\name{synthetic_field_hcc}
\alias{synthetic_field_hcc}
\title{Synthetic Hard Coral Cover Field}
\usage{
synthetic_field_hcc(
  spatial_grid,
  all_effects_df,
  baseline_sample_hcc,
  spde,
  config_sp
)
}
\arguments{
\item{spatial_grid}{An \code{sfc_POINT} object representing the full spatial grid.}

\item{all_effects_df}{A data frame containing the effects of the year.}

\item{baseline_sample_hcc}{A data frame containing the baseline hard coral cover sample.}

\item{spde}{A list containing the SPDE mesh, SPDE object, precision matrix \code{Q},
and projection matrix \code{A}.}

\item{config_sp}{A list containing simulation config_spuration parameters.}
}
\value{
A list with:
\itemize{
\item \code{all_effects_hcc} – matrix of combined baseline and yearly effects
\item \code{all_pts_sample_hcc} – projected synthetic field onto the spatial grid
\item \code{all_pts_effects_hcc} – long-format data frame suitable for plotting
}
}
\description{
Synthetic Hard Coral Cover Field
}
\details{
Creates a broad-scale synthetic field of hard coral cover by combining
the baseline field with the effects of the year. All effects are assumed
to be on the link scale (logit) and projected onto the SPDE grid.
}
\author{
Murray
}
